+++ docs/src/pages/components/layout/AccordionDemo.tsx
@@ -153,7 +153,6 @@ export default function AccordionDemoPage() {
           </Tabs.Panel>
         </Tabs>
       </Stack>
-      
     </Surface>
   );
 }